paydirt-web/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
pyyaml==6.0.1
